페이즈0 백업/requirements.txt
requests
supabase
pytz
tenacity
groq
google-genai